Update _yaourt
I have the next function in .zshrc:
      pacman () {
        /usr/bin/sudo /usr/bin/pacman $* && echo "$*" | grep -q "S\|R" && rehash
      }

And I have the next line in /etc/yaourtrc:
      PACMAN="/usr/bin/pacman"

But if I try to get autocomplite for `yaourt -Qs` (for example) it requires the sudo password. With this patch path to `pacman` will be read from the configuration file and this command will not require the password.