Support newer versions of OpenSSL
With OpenSSL 1.1.0e (likely earlier as well) the `list-*` commands are
no longer supported causing completion to fail for subcommands and
arguments of digest and cipher subcommands.

Instead use the output of `openssl help`, removing the section headers
and splitting the list of commands which that displays.  Earlier
versions don't actually have a `help` subcommand but they did already
produce the same help output for unknown subcommands, just preceded with
an error message about the invalid command. This error message is
removed along with the section headers.

One use of the obsolete `list-cipher-algorithms` subcommand remains, but
it is in the completion for the `genpkey` subcommand which is also
obsolete.