Rename go completion to not collide with zsh file.
There is a _go file in the zsh functions directory that would be found
first if you use OMZ with zsh-completions.  This is an ancient file in
zsh and does not complete for the go executable either, but it is still
there on typical zsh distributions.

This change makes it such that zsh/OMZ/zsh-completions can work in
harmony, if one so chooses.