Fix kak_sessions not sending _values enough arguments
Using zsh 5.4.2, I am occasionally running into an error "_values:compvalues:11: not enough arguments". When it doesn't occur, instead, one of the kakoune sessions does not appear as part of autocompletion.

This fixes both problems for me by providing a description as described in http://zsh.sourceforge.net/Doc/Release/Completion-System.html#Completion-System. "The first non-option argument is used as a string to print as a description before listing the values. "